Add header and add background image
Added a div element with id header at the top of the html body. This
will be used to add a website title and an eventual button to change
the number of pixels on the board. I also gave the board a specific
width and height to keep it from stretching.