Add header, instructions for game, and button
First off, I removed the container element because there were too many
elements which were causing problems with the flex-box positioning.
Second, I added a header section directly in html which includes a title
and a slogan. Below the header section, I added an instructions section
which includes instructions for the etch-a-sketch game, and a non-
stylized button which will change the pixel size of the board pixels in
the next commmit. I also added a new font to the game folder.