Apply event delegation to fix the event listener
The bug related to the mouse hover event listener has been fixed. I made
the etch-a-sketch element a statically created element in the html, and
I added the event listener to it rather than the dynamically created
children with class "square". I also used a conditional statement to
have the event performed on the individual squares rather than the whole
board. This is taking advantage of "event propagation".