Complete extra credit
The functionality of the etch-a-sketch game is now 100% complete. This
extra credit involved darkening the square, after it was assigned a
random colour, by 10% after each following pass of the cursor. This
involved the initialization of a new array and temporary variables to
store in memory the original colour value, and the use of regex
expressions to receive and modify colour values via temporary variables.
Finally, a conditional statement was used in the event listener to
determine whether it was the first time passing over a square or not.